:recycle: テキストフィールドをフォームエリアから分離。
