+++ components/layouts/FileFormUploadArea.tsx
@@ -3,45 +3,6 @@ import { SentimentSatisfied } from '@material-ui/icons';
 import React, { useState } from 'react';
 
 const FileFormUPloadArea = () => {
-  const handleErrorMessage = () => {
-    if (!isSubmit) {
-      return '';
-    }
-    if (sampleValue === '') {
-      return '入力してください。';
-    }
-    return '';
-  };
-
-  const infoMessage = (infoMessages: string[] = []) => {
-    return infoMessages?.length ? (
-      <div>
-        {infoMessages.map((infoMaggage) => (
-          <p>{infoMaggage}</p>
-        ))}
-      </div>
-    ) : (
-      ''
-    );
-  };
-
-  const checkError = () => {};
-
-  const errorMessage = (errorMessages: string[] = []) => {
-    return errorMessages?.length ? (
-      <div>
-        {errorMessages.map((errorMessage) => (
-          <p>{errorMessage}</p>
-        ))}
-      </div>
-    ) : (
-      ''
-    );
-  };
-
-  const [sampleValue, setSampleValue] = useState('');
-  const [isSubmit, setIsSubmit] = useState(false);
-
   return (
     <>
       <form
@@ -55,30 +16,8 @@ const FileFormUPloadArea = () => {
       >
         <div className="p-5">
           <ul>
-            <li>
-              <TextField
-                id="outlined-basic"
-                name="name1"
-                label="サンプル12"
-                variant="outlined"
-                error={isSubmit && true}
-                helperText={infoMessage()}
-                onChange={(e) => {
-                  setSampleValue(e.currentTarget.value);
-                  if (isSubmit) checkError();
-                }}
-              />
-              {isSubmit.toString()}
-            </li>
-
-            <li>
-              <TextField
-                id="outlined-basic2"
-                name="name2"
-                label="サンプル2"
-                variant="outlined"
-              />
-            </li>
+            <li></li>
+            <li></li>
             <li></li>
             <li></li>
             <li></li>
